+++ blivet/devices/md.py
@@ -433,11 +433,6 @@ class MDRaidArrayDevice(ContainerDevice, RaidDevice):
         return list(self.parents)
 
     @property
-    @util.deprecated('1.11', "Use 'members' instead.")
-    def devices(self):
-        return self.members
-
-    @property
     def complete(self):
         """ An MDRaidArrayDevice is complete if it has at least as many
             component devices as its count of active devices.
@@ -686,11 +681,6 @@ class MDBiosRaidArrayDevice(MDRaidArrayDevice):
         # is a container and its actual member devices are the
         # container's parents.
         return self.parents[0].members
-
-    @property
-    @util.deprecated('1.11', "Use 'members' instead.")
-    def devices(self):
-        return self.parents[0].devices
 
     @property
     def total_devices(self):